+++ src/include/migraphx/marker.hpp
@@ -26,8 +26,8 @@ inline namespace MIGRAPHX_INLINE_NS {
  * struct marker
  * {
  *      void mark(std::string st) ;
- *      uint64_t range_start(std::string st) ;
- *      void range_stop(uint64_t range_num) ;
+ *      size_t range_start(std::string st) ;
+ *      void range_stop(size_t range_num) ;
  *      void trace_ins_start(std::string st) ;
  *      void trace_ins_end() ;
  * };
@@ -103,13 +103,13 @@ struct marker
         (*this).private_detail_te_get_handle().mark(std::move(st));
     }
 
-    uint64_t range_start(std::string st)
+    size_t range_start(std::string st)
     {
         assert((*this).private_detail_te_handle_mem_var);
         return (*this).private_detail_te_get_handle().range_start(std::move(st));
     }
 
-    void range_stop(uint64_t range_num)
+    void range_stop(size_t range_num)
     {
         assert((*this).private_detail_te_handle_mem_var);
         (*this).private_detail_te_get_handle().range_stop(std::move(range_num));
@@ -141,8 +141,8 @@ struct marker
         virtual const std::type_info& type() const                                = 0;
 
         virtual void mark(std::string st)            = 0;
-        virtual uint64_t range_start(std::string st) = 0;
-        virtual void range_stop(uint64_t range_num)  = 0;
+        virtual size_t range_start(std::string st)   = 0;
+        virtual void range_stop(size_t range_num)    = 0;
         virtual void trace_ins_start(std::string st) = 0;
         virtual void trace_ins_end()                 = 0;
     };
@@ -177,13 +177,13 @@ struct marker
 
         void mark(std::string st) override { private_detail_te_value.mark(std::move(st)); }
 
-        uint64_t range_start(std::string st) override
+        size_t range_start(std::string st) override
         {
 
             return private_detail_te_value.range_start(std::move(st));
         }
 
-        void range_stop(uint64_t range_num) override
+        void range_stop(size_t range_num) override
         {
 
             private_detail_te_value.range_stop(std::move(range_num));
