+++ tools/include/marker.hpp
@@ -23,8 +23,8 @@ inline namespace MIGRAPHX_INLINE_NS {
 <%
 interface('marker',
            virtual('mark', st = 'std::string', returns = 'void'),
-           virtual('range_start', st = 'std::string', returns = 'uint64_t'),
-           virtual('range_stop', range_num = 'uint64_t', returns = 'void'),
+           virtual('range_start', st = 'std::string', returns = 'size_t'),
+           virtual('range_stop', range_num = 'size_t', returns = 'void'),
            virtual('trace_ins_start', st = 'std::string', returns = 'void'),
            virtual('trace_ins_end', returns = 'void')
         ) %>
